update app insights and add API method for clearing telemetry

+++ Telimena/Telimena.WebApp.Infrastructure/Repository/Implementation/ProgramRepository.cs
@@ -6,6 +6,7 @@ using System.Linq.Expressions;
 using System.Runtime.Caching;
 using System.Security.Principal;
 using System.Threading.Tasks;
+using Telimena.WebApp.Core.DTO;
 using Telimena.WebApp.Core.DTO.MappableToClient;
 using Telimena.WebApp.Core.Interfaces;
 using Telimena.WebApp.Core.Models;
@@ -98,7 +99,7 @@ namespace Telimena.WebApp.Infrastructure.Repository.Implementation
         }
         public void Remove(Program program)
         {
-            this.ClearTelemetryData(program);
+            this.ClearTelemetryData(program, null);
             this.telemetryContext.TelemetryRootObjects.RemoveRange(this.telemetryContext.TelemetryRootObjects.Where(x => x.ProgramId == program.Id));
 
             this.portalContext.Versions.RemoveRange(this.portalContext.Versions.Where(x => x.ProgramAssembly.ProgramId == program.Id));
@@ -109,45 +110,52 @@ namespace Telimena.WebApp.Infrastructure.Repository.Implementation
             this.portalContext.Programs.Remove(program);
         }
 
-        private void ClearTelemetryData(Program program)
+
+        public void ClearTelemetryData(Program program, TelemetryItemTypes? type)
         {
             IQueryable<Core.Models.Telemetry.ClientAppUser> eventUsers = this.telemetryContext.EventTelemetrySummaries.Where(x => x.Event.ProgramId == program.Id).Select(x => x.ClientAppUser);
             IQueryable<Core.Models.Telemetry.ClientAppUser> viewUsers = this.telemetryContext.ViewTelemetrySummaries.Where(x => x.View.ProgramId == program.Id).Select(x => x.ClientAppUser);
 
             List<int> userIds = eventUsers.Select(x => x.Id).Concat(viewUsers.Select(x => x.Id)).Distinct().ToList();
 
+            if (type == null || type == TelemetryItemTypes.View)
+            {
+                this.telemetryContext.ViewTelemetryUnits.RemoveRange(this.telemetryContext.ViewTelemetryUnits.Where(x => x.TelemetryDetail.TelemetrySummary.View.ProgramId == program.Id));
+                this.telemetryContext.ViewTelemetryDetails.RemoveRange(this.telemetryContext.ViewTelemetryDetails.Where(x => x.TelemetrySummary.View.ProgramId == program.Id));
+                this.telemetryContext.ViewTelemetrySummaries.RemoveRange(this.telemetryContext.ViewTelemetrySummaries.Where(x => x.View.ProgramId == program.Id));
+                this.telemetryContext.Views.RemoveRange(this.telemetryContext.Views.Where(x => x.ProgramId == program.Id));
+            }
 
-            this.telemetryContext.ViewTelemetryUnits.RemoveRange(this.telemetryContext.ViewTelemetryUnits.Where(x => x.TelemetryDetail.TelemetrySummary.View.ProgramId == program.Id));
-            this.telemetryContext.ViewTelemetryDetails.RemoveRange(this.telemetryContext.ViewTelemetryDetails.Where(x => x.TelemetrySummary.View.ProgramId == program.Id));
-            this.telemetryContext.ViewTelemetrySummaries.RemoveRange(this.telemetryContext.ViewTelemetrySummaries.Where(x => x.View.ProgramId == program.Id));
-            this.telemetryContext.Views.RemoveRange(this.telemetryContext.Views.Where(x => x.ProgramId == program.Id));
 
-            this.telemetryContext.EventTelemetryUnits.RemoveRange(this.telemetryContext.EventTelemetryUnits.Where(x => x.TelemetryDetail.TelemetrySummary.Event.ProgramId == program.Id));
-            this.telemetryContext.EventTelemetryDetails.RemoveRange(this.telemetryContext.EventTelemetryDetails.Where(x => x.TelemetrySummary.Event.ProgramId == program.Id));
-            this.telemetryContext.EventTelemetrySummaries.RemoveRange(this.telemetryContext.EventTelemetrySummaries.Where(x => x.Event.ProgramId == program.Id));
-            this.telemetryContext.Events.RemoveRange(this.telemetryContext.Events.Where(x => x.ProgramId == program.Id));
+            if (type == null || type == TelemetryItemTypes.Event)
+            {
+                this.telemetryContext.EventTelemetryUnits.RemoveRange(this.telemetryContext.EventTelemetryUnits.Where(x => x.TelemetryDetail.TelemetrySummary.Event.ProgramId == program.Id));
+                this.telemetryContext.EventTelemetryDetails.RemoveRange(this.telemetryContext.EventTelemetryDetails.Where(x => x.TelemetrySummary.Event.ProgramId == program.Id));
+                this.telemetryContext.EventTelemetrySummaries.RemoveRange(this.telemetryContext.EventTelemetrySummaries.Where(x => x.Event.ProgramId == program.Id));
+                this.telemetryContext.Events.RemoveRange(this.telemetryContext.Events.Where(x => x.ProgramId == program.Id));
+            }
 
-            this.telemetryContext.LogMessages.RemoveRange(this.telemetryContext.LogMessages.Where(x => x.ProgramId == program.Id));
+            if (type == null || type == TelemetryItemTypes.LogMessage)
+            {
+                this.telemetryContext.LogMessages.RemoveRange(this.telemetryContext.LogMessages.Where(x => x.ProgramId == program.Id));
+            }
 
-            this.telemetryContext.ExceptionTelemetryUnits.RemoveRange(this.telemetryContext.ExceptionTelemetryUnits.Where(x => x.ExceptionInfo.ProgramId == program.Id));
-            this.telemetryContext.Exceptions.RemoveRange(this.telemetryContext.Exceptions.Where(x => x.ProgramId == program.Id));
+            if (type == null || type == TelemetryItemTypes.Exception)
+            {
+                this.telemetryContext.ExceptionTelemetryUnits.RemoveRange(this.telemetryContext.ExceptionTelemetryUnits.Where(x => x.ExceptionInfo.ProgramId == program.Id));
+                this.telemetryContext.Exceptions.RemoveRange(this.telemetryContext.Exceptions.Where(x => x.ProgramId == program.Id));
+            }
 
             foreach (var userId in userIds)
             {
-                var isInUseEvents  = this.telemetryContext.EventTelemetrySummaries.Where(x=>x.Event.ProgramId != program.Id).Any(x => x.ClientAppUserId == userId);
-                var isInUseViews  = this.telemetryContext.ViewTelemetrySummaries.Where(x => x.View.ProgramId != program.Id).Any(x => x.ClientAppUserId == userId);
+                var isInUseEvents = this.telemetryContext.EventTelemetrySummaries.Where(x => x.Event.ProgramId != program.Id).Any(x => x.ClientAppUserId == userId);
+                var isInUseViews = this.telemetryContext.ViewTelemetrySummaries.Where(x => x.View.ProgramId != program.Id).Any(x => x.ClientAppUserId == userId);
 
                 if (!isInUseEvents && !isInUseViews)
                 {
                     this.telemetryContext.AppUsers.RemoveRange(this.telemetryContext.AppUsers.Where(u => u.Id == userId));
                 }
             }
-
-        }
-
-        public void ClearTelemetryAllData(Program program)
-        {
-            this.ClearTelemetryData(program);
         }
 
         public async Task<IEnumerable<Program>> GetAsync(Expression<Func<Program, bool>> filter = null
